Added new DB user for web app and Search filters
Added a new user with only read/write perms. This should be used instead of the root in the backend. Set the password when running the sql script then set that password in app.py

In addition, added rating filtering to location search, including tags system.

(Will be updated to look nicer later)